Fix handling of very long MP3 and M4A files.

+++ components/decoder/mad/mad.h
@@ -35,8 +35,8 @@ namespace BoCA
 
 			Track			*infoTrack;
 
-			Int			 numBytes;
-			Int			 numFrames;
+			Int64			 numBytes;
+			Int64			 numFrames;
 
 			Int			 delaySamples;
 			Int			 padSamples;
