+++ components/decoder/mpg123/mpg123.h
@@ -26,8 +26,8 @@ namespace BoCA
 
 			Buffer<unsigned char>	 samples;
 
-			Int			 numBytes;
-			Int			 numFrames;
+			Int64			 numBytes;
+			Int64			 numFrames;
 
 			Int			 delaySamples;
 			Int			 padSamples;
