+++ components/encoder/faac/faac.h
@@ -1,5 +1,5 @@
  /* BoCA - BonkEnc Component Architecture
-  * Copyright (C) 2007-2018 Robert Kausch <robert.kausch@freac.org>
+  * Copyright (C) 2007-2019 Robert Kausch <robert.kausch@freac.org>
   *
   * This program is free software; you can redistribute it and/or
   * modify it under the terms of the GNU General Public License as
@@ -38,7 +38,7 @@ namespace BoCA
 			Int			 blockSize;
 			Int			 overlap;
 
-			Int			 totalSamples;
+			Int64			 totalSamples;
 
 			Buffer<int16_t>		 samplesBuffer;
