+++ components/encoder/voaacenc/voaacenc.h
@@ -1,5 +1,5 @@
  /* BoCA - BonkEnc Component Architecture
-  * Copyright (C) 2007-2017 Robert Kausch <robert.kausch@freac.org>
+  * Copyright (C) 2007-2019 Robert Kausch <robert.kausch@freac.org>
   *
   * This program is free software; you can redistribute it and/or
   * modify it under the terms of the GNU General Public License as
@@ -36,7 +36,7 @@ namespace BoCA
 
 			Int			 frameSize;
 
-			Int			 totalSamples;
+			Int64			 totalSamples;
 			Int			 delaySamples;
 
 			Buffer<unsigned char>	 outBuffer;
